rdp-pipewire-stream: Add signal to indicate that video size changed
In multimonitor setups, gnome-remote-desktop has to properly map all
surfaces to the graphics output buffer ADM element.
As a result, it is necessary to know that the video size of the stream
for a surface was correctly resized.
In addition to that, some sort of feedback is required for the layout
manager to be able to avoid a race, when a lot of new monitor configs
are received from the remote desktop client.

So, add this signal to help solving these challenges.