rdp: Use virtual monitor size instead of surface size for layout updates
While the surface size and virtual monitor size must match, it just
leads to confusion here, when reading the code.
In addition to that, using the surface size for virtual monitors would
only work, if the surface was already resized, which will not always be
the case, when multiple monitors are being used.