session-rdp: Add API to notify new desktop size
In multimonitor setups, the desktop area does not only include the area
of the current surface, but from all surfaces.
Since the current surface is unaware of this data, but the layout
manager, add an API to notify the new desktop size.
This API will then be used in the next commits by the layout manager.