rdp-graphics-pipeline: Always reset RFX context
Resetting the RFX context has two reasons:
1. To ensure that the size of the input frame matches the size in the
   context.
2. To reset the state of the RFX encoder to include an RFX header again
   in RFX messages.

Reason 2 is only relevant in the legacy path, but not in the graphics
pipeline, since RFX Progressive uses its own message headers.
Setting the width and height of the input frame is a quick operation.
So, remove any condition preventing from doing it.