rdp-layout-manager: Add API to invalidate surfaces
This API will later be used to invalidate all surfaces, when the
graphics pipeline has been reset.
Don't invalidate surfaces, however, when the layout manager is in a
state of a layout transition, since not all surfaces may be well defined
yet.
This is though not a problem since all surfaces are invalidated anyway,
when the layout changes and therefore the invalidation call is not lost.