session-rdp: Adapt to the new invalidate_surface API
Currently, when all surfaces need to be invalidated,
gnome-remote-desktop sets the RDP_PEER_ALL_SURFACES_INVALID flag.
But since the layout manager now manages all surfaces, that are mapped
to outputs, this task needs to be forwarded to it now.

So, when invalidating all surfaces due to a reset of the graphics
pipeline, simply forward this task to the layout manager.
Since the RDP_PEER_ALL_SURFACES_INVALID flag is now superfluous, remove
it.