rdp-pipewire-stream: Fix race condition when stream is destroyed
The render handling always happens in the graphics thread.
But destroying a PipeWire stream happens in the main thread.
While the PipeWire stream needs to be destroyed in the main thread, the
render sources in the PipeWire stream class still access the stream in
their callbacks.
This is an issue, that can happen in multimonitor scenarios, where a
virtual monitor is destroyed during the session.
It cannot happen currently, since only single monitor sessions are
allowed and there PipeWire streams for monitors are only destroyed after
the graphics thread stopped.

To get rid of this issue, simply introduce a stream context for each
render source.
This render source contains a mutex and a reference to the PipeWire
stream class.
Before the render sources access the stream class, they will lock the
stream mutex and check whether the stream is not NULL.
If it is NULL, the render handling is not performed, and the GSource
destroyed.

When destroying the PipeWire class, the stream mutex is locked too, and
the stream simply set to NULL.
This handles two situations:
1. There is no more a use-after-free situation with the stream class
   instance.
2. There is no roundtrip through the graphics thread needed. Such
   roundtrip would introduce its own problems here (this can include
   potential deadlocks).