rdp-layout-manager: Fix render sources not being triggered when wanted
When the layout manager finishes a layout transition, it triggers all
render sources. This happens in the START_RENDERING state.
However, maybe_trigger_render_sources() requires the layout manager to
be in the AWAIT_CONFIG config state, otherwise render sources are not
triggered.
If a frame for a surface was already queued at this point, it is
therefore not rendered.

To fix this issue, simply run maybe_trigger_render_sources() after the
state transition to AWAIT_CONFIG, since then the layout transition
completed.

The issue itself may only occur seldom, since the layout transition
happens relatively fast.