settings: Modify how the server-certificate and private-key are stored
Instead of just storing the path of the server certificate and the
private key, also store their content.

The content of the server certificate and the private key will in future
commits not just be used by the RDP server itself, but also for the
handover process.
While the path of the server certificate and the private key is
primarily used by the different settings backends, the content of both
files will in future commits be fetched by the handover daemon from the
system daemon to use the same server certificate and private key, that
the system daemon uses.