Introduce daemon-system
This daemon is running as a system service.

It is not in charge of displaying any session.

It only listens to the RDP port and on a new connection it creates a
RemoteClient and uses the gdm interface RemoteDisplayFactory to create a
RemoteDisplay giving the id of this RemoteClient (aka remote_id).

There's one RemoteClient per one RDP client.

Each RemoteDisplay is "tagged" with one RemoteClient id (remote_id).