daemon-system: Register handover iface when a new RemoteDisplay is added
The system daemon connects to the ObjectManagerServer of gdm to know
when a new display object is added.

When a display interface added, it checks:
- If it is a RemoteDisplay
- If its remote_id corresponds to an id of any of the stored
  RemoteClients

If both conditions are true, a handover interface is registered with the
session_id of the RemoteDisplay
(i.e. /org/gnome/RemoteDesktop/Rdp/Handovers/session<session_id>) and
that interface proxy is stored on the RemoteClient, which has the
looked up remote_id.

The handover interface will be implemented in next commits.