daemon-handover: Logout when RDP client disconnects or handover fails
There are three possible situations, where a logout should happen:

1. The RDP client disconnected.
2. Calling "StartHandover" method returned an error from daemon-system.
3. The "StartHandover" method is called successfully, but after a while,
   the "TakeClientReady" signal isn't received.

The RDP client mstsc disconnects and reconnects a second time on
redirection, when using RDSTLS as security method.
So use the "incoming-new-connection" signal to make sure to not logout
on the first disconnection, but to logout when the client finished
authenticating ("incoming-new-connection" is emitted on the PostConnect
callback).

In the future, user sessions won't be logged out, making it possible to
reuse them in a new connection.