settings-user: Use default RDP and VNC ports for user daemon
Originally, it was intended to reserve the default RDP and VNC ports for
the system daemon, that will be introduced with the headless session
merge requests.
However, it is time-wise very unlikely, that the system daemon is
integrated in GNOME 45.
Thus revert the port change of the user daemon to use the default RDP
and VNC ports in the user daemon again.