vnc-pipewire-stream: Do not prematurely clear PipeWire stream
With the now implemented blocking mechanism in the remove_buffer
callback, PipeWire won't destroy buffers anymore, as long they are still
in use.
However, in order to queue a PipeWire buffer again, the PipeWire stream
is needed. But when destroying the stream, the stream is set to NULL
before it is destroyed.
This is problematic on the 42.x branch, as on that branch PipeWire
buffers can still be dequeued, while the EGL thread finishes processing
of the already dequeued PipeWire buffers.
As a result, these newly dequeued buffers trigger a crash, when queueing
these buffers again, as the respective PipeWire stream is already NULL.

Fix this issue, by not setting the PipeWire to NULL, when destroying it.