vnc-pipewire-stream: Don't remove buffer while dequeueing it
When a PipeWire stream starts, then the PipeWire realtime thread will
run without a break.
This is also the case, when a PipeWire buffer is about to be destroyed:
The main thread, where the PipeWire loop runs can destroy a PipeWire
buffer, while the PipeWire realtime thread processes its contents.
As a result, a use-after-free situation can happen.

To avoid this issue, lock the dequeue mutex before destroying a PipeWire
buffer. This ensures, that either a PipeWire buffer is processed before
it is destroyed, or that a buffer lock can be acquired before the buffer
is attempted to be destroyed.
In this latter case here, the main thread would block until the acquired
buffer lock is released again.