rdp: Also include DRM format modifiers when resizing PipeWire stream
When resizing the PipeWire stream, all PipeWire buffers need to be
recreated. For this, the supported DRM format modifiers need to be
advertised to mutter again, as mutter checks the supported modifiers on
the stream consumer side not just on the initial creation, but also on
every resize.
As a result, mutter will with gnome-remote-desktops current behaviour
not use dma-bufs again, when the stream is resized, because
gnome-remote-desktop only advertises the supported DRM format modifiers
on the initial stream creation.

To be able to use dma-bufs even after a resize, advertise all supported
DRM format modifiers again, when preparing the parameters for the
resized stream.