session: Also clear related EI regions when disposing EI abs pointer
When the EI pointer device is removed, gnome-remote-desktop currently
still happily continues using any associated regions, because there are
still references to them and their EI devices.
In addition to that, regions are currently not removed, unless the
session is stopped resulting in a waste of memory.

Fix this issue by properly disposing these regions, when removing the
respective EI device.