rdp-dsp: Add API to retrieve frames-per-packet value
Certain encoders require specific packet sizes, when submitting frames.
This, for example, applies to fdk-aac and the Opus codec. However,
these sizes are not necessarily the same across those encoders.
So, add an API to retrieve those packet sizes. This API will then be
used in the following commits to be able to add support for the Opus
codec, while retaining support for fdk-aac.