rdp-layout-manager: Always properly release stream id
When a record call for a (virtual) monitor was done, but the call did
not return yet, then the associated stream id won't be released.
This situation can only occur, when the record call directly happens
before the session is stopped.
Since the stream ids are cleared anyway on exit, this change won't have
a huge effect. However, logic-wise, it does not make sense to only
release the stream id, when an actual stream was acquired, since the
stream id is acquired before any record call happens.