rdp-surface: Move rendering functionality to separate class
The RDP surface itself only describes an area, that will later be used
for an output or a window.
The rendering functionality works on that surface, but it is not a
substantial part of the surface itself. So, move it into a separate
class.
This also eases up making the RDP surface opaque later, as a few public
members are removed.