stream: Stop stream before destroying it
When destroying individual streams, mutter won't know, that the stream
consumer is actually gone.
For virtual monitors, this is especially problematic, because the
respective virtual monitor of the stream would linger around.

Stopping the stream via the new Stop() API signals mutter to destroy the
whole stream, and with that any associated virtual monitors.
So, do that before destroying the stream on gnome-remote-desktops side.