session: Always properly clear regions hash table
When a session was created, but not successfully initialized,
gnome-remote-desktop crashes, because the regions hash table, that is
created upon session creation, is only cleared, when the session is
stopped and thus an assertion is hit here.
The reason why this assertion hit only happens when the session
initialization fails, is that, that in this situation, the session is
not stopped, but directly cleared, and so this assertion triggers.

Fix this issue, by manually clearing this hash table in finalize() too.

Fixes: https://errors.ubuntu.com/problem/25bcb4cc3c4d388e4dc1f626d015b15d63712842