daemon-handover: Gracefully shutdown stray gdm handover daemon
To be able to hand over connections to a daemon-handover, the daemon-handover
needs to be automatically started, when the respective
gnome-session starts for a user.
The exception here is when the daemon-handover is started at a local gdm
session. In that case it is unused and just unnecessarily occupies memory.

So, try to detect whether the current running daemon-handover instance
is that unused one, and if it is, just shut down the process gracefully.

TODO: Make sure GDM sets the logind remote property to true in sessions with a
RemoteDisplay.