stream: Do not crash when stream parameters could not be retrieved
g_variant_lookup() requires the dictionary (here: parameters) to be
non-NULL, otherwise a crash occurs.
So, check that before attempting to look up the mapping-id.

While it is unclear why parameters is here NULL in the first place,
sanitizing input (here: by checking against NULL), that comes from the
outside, is a necessary action.

Fixes: https://errors.ubuntu.com/problem/7ce9f4812422da55e7cb4c1a1577d666d1fc2e67