session: Add struct for absolute pointer motion events
Absolute pointer motion events may be based upon an input region, that
has a different size than the EI region to which the position is
transformed to.
To be able to translate between these regions, add this struct,
containing both the absolute pointer position and the size of the region
to which this position relates to.