session-rdp: Require clients to support fastpath output
gnome-remote-desktop already makes use of surface commands, which
require support for fastpath output, so require the client to support
it too.
This is safe to use, since MS Windows RDS already requires the same on
Microsoft RDP 7.1 servers and on later versions.
As a result, Microsofts clients and FreeRDP-based clients already
support the fastpath output (it is itself already ancient) and this
check primarily aims to ensure, that gnome-remote-desktops behaviour is
conform to the protocol.