rdp: Adapt to cursor renderer
With the cursor renderer now been implemented, adapt the cursor
rendering handling to it.
For this, remove all cursor rendering actions in the PipeWire stream
classes and simply just submit updates to the cursor renderer.

In the session, create the cursor renderer, when the remote desktop
session is ready.
The cursor renderer cannot be created earlier, because it relies on the
pointer cache size to be negotiated.