session-rdp: Prepare for FreeRDP3
In FreeRDP2, the RDP peer instance and RDP context instance share
several members, such as the settings instance or the update instance.
FreeRDP3 removes these members from the peer instance, so several getter
operations need to be adjusted.

Separate this work here out of the actual FreeRDP3 port to not only
reduce the overall line difference of the port, but also to make
rebasing easier, when commits were developed on top of the FreeRDP3
version of gnome-remote-desktop.
Functionally, nothing changes here.