fixup! daemon-system: Register handover iface when a new RemoteDisplay is added
In theory, the object manager objects can come with our without the
needed interface on them, so it's best to handle both cases.

This commit listens for object-added/object-removed to know when new
display objects are added, and also listens for notify::remote-display
(which is sugar around interface-added/interface-removed) to know when
the objects get the needed interface.