rdp-server: Add the ability to "find a port dynamically"
This is the behaviour when 'negotiate-port' setting
is enabled.

If there is more than one user running an RDP server,
or if there is a system gnome-remote-desktop instance
alongside a user gnome-remote-desktop, or if there is
a headless gnome-remote-desktop running while a user
gnome-remote-desktop instance is running there can easily
be contention between instances when binding to ports.

Now the settings-user uses this as a default.