rdp-server: Allow 0 port to mean "find one dynamically"
If there is more than one user running an RDP server,
or if there is a system gnome-remote-desktop instance
alongside a user gnome-remote-desktop, or if there is
a headless gnome-remote-desktop running while a user
gnome-remote-desktop instance is running there can easily
be contention between instances when binding to port 3389.

This commit helps resolve the contention by allowing a port
of 0 to mean "figure it out dynamically".