session-rdp: Make use of FreeRDP settings getter and setter methods
Currently, gnome-remote-desktop accesses the FreeRDP-settings attributes
directly. However, this access will be removed in future FreeRDP
versions, as the FreeRDP settings structure is intended to be opaque.
As a replacement, FreeRDP offers getter and setter methods for the
respective settings.
These getter and setter methods are already available in FreeRDP2, so
use these now already, as a preparatory step for these future versions.