settings: Add system backend
This setting backend will be used by the system daemon. It hardcodes
the enabled, view-only and screen-share-mode settings.
The RDP server-certificate- and private-key- settings are stored in a
keyfile called grd.conf.

This settings backend will write the key-file when the respective
settings are set. It will also read the key-file again, when it is
written/saved.

To avoid ending up in an infinite loop with reading and writing, use
the g_signal_handlers_block_by_func and
g_signal_handlers_unblock_by_func functions.