settings-user: Set RDP/VNC ports at gsettings
The default ports are still 3389 for the RDP server and 5900 for the VNC
server.

However, instead of being hardcoded, now they are at settings (gsettings
in the settings-user backend).

It is still possible to override the port if running the daemon with the
--rdp-port/--vnc-port options.

This change adds more flexibility to decide which RDP/VNC port will use each
session and prepares the possibility of choosing a port automatically.