rdp-audio-playback: Add support for the Opus codec
With the encode handling of the Opus codec now added, hook it now up.
One major difference to the other encode paths is that the audio data is
sampled at 48000Hz, instead of 44100Hz. The reason for this is that Opus
does not support 44100Hz, however, it supports 48000Hz as sample rate.

Also, increase the minimum idling time from 5 seconds to 10 seconds,
when the Opus codec is used.
The idling time determines the earliest point in time, when an audio
stream is going to be unlocked. The reason for this time is to ensure,
that all sent samples are flushed (played), before the overall audio
stream is paused. This avoids playing old samples, when the remote
desktop client is closed, but no new audio samples didn't play for some
time. The idling time also ensures no unnecessary problems with music,
when it contains short pauses.
For audio data at 48000Hz (required for Opus), the default of 5 seconds
is not enough. 10 seconds, however, appears to be enough in testing for
the idling time, so increase the idling time, when the Opus codec is
used.