rdp-surface-renderer: Add method to submit new buffers for rendering
Currently, there are two render sources per surface in the RDP backend:
1. In the PipeWire stream class
2. In the surface renderer class

Render source 1 is used, when a new frame is ready, while render source
2 is used in all other cases, e.g. when rendering was inhibited or
suspended and in the meantime a new frame was submitted.

The render source in the PipeWire stream class is actually superfluous.
It also makes things more complicated, since the surface and the surface
renderer outlive the PipeWire stream class. This is necessary, as the
render source in the PipeWire stream class still needs to run in the
graphics thread.
In preparation to get rid of the render source in the PipeWire stream
class, add a method to the surface renderer to submit new buffers.
This allows in the next to get rid of this superfluous render source.