rdp-pipewire-stream: Submit buffers for rendering to surface renderer
With the method to submit buffers to the surface renderer now been
implemented, use it now in the PipeWire stream class to submit buffers
to the surface renderer.
The surface renderer automatically picks these buffers up, when the time
is right.
The old render source in the PipeWire stream class is now unnecessary,
so remove it.
This also gets rid of the StreamContext, which was previously necessary
to avoid race conditions, when the PipeWire stream instance was
destroyed. Overall, this simplifies the rendering handling a bit.