Add headless systemd service
This allows one to start a headless user daemon instance.

To make this integrate better with systemd, use different D-Bus bus
names depending on the runtime mode, so that each systemd user service
can have its own unique bus name. Not doing this will result in the
following error:

> Two services allocated for the same bus name org.gnome.RemoteDesktop, refusing operation.

The headless service files unit is marked as Conflicts= with the
non-headless user service, to allow both being enabled, but that could
be reconsidered in the future.