rdp: Add renderer class
Currently, all rendering actions happen in the session-rdp class. This
was fine for a long time, since gnome-remote-desktop only needed to
handle one monitor.
But even with just one monitor, the session-rdp class is pretty
cluttered.

To take care of this (and of future challenges), introduce a new
renderer class. This class will manage all render actions, such as
inhibition of rendering, or anything else, that is related to rendering,
but not specific to certain surfaces.
For now, add handling for the graphics thread, so it can be moved into
this class in the next commit. And also add the surface disposal source
into this class, which currently resides in the layout manager, where it
should not be.
In future commits, more handling is moved into the renderer, such as the
rendering inhibition or the graphics reset handling.