rdp: Adapt to RDP renderer
With the base class of the renderer now being available, adapt to it.
Use the acquire_surface() and release_surface() surface methods in the
layout manager to acquire and release surfaces instead of managing these
actions in the layout manager.
Also, use the graphics thread (and context) in the renderer, instead of
the equivalent ones in session-rdp.