rdp-audio-playback: Protect against unexpected Training Confirm PDUs
Before the Training Confirm PDU is received, the audio formats are
exchanged. Only after that, gnome-remote-desktop can prepare the client
format.
In addition to that, the Training Confirm PDU MUST only be sent by the
client, when the server sent a Training PDU.
To anticipate a protocol violation and an unexpected situation, simply
tear down the virtual channel, when a Training Confirm PDU was received,
but no audio formats.