rdp-dsp: Stop making fdk-aac optional
When support for fdk-aac was added, the patent situation was unclear.
However, this has been resolved in the meantime: gnome-remote-desktop
makes use of the low-complexity profile, which is perfectly suited for
the bitrate, that gnome-remote-desktop uses. The patents of this profile
expired half a decade ago, so there is no reason to have this dependency
optional.
Distributions, like Fedora, that still have concerns already patch out
any functionality, that the LC profile does not depend on (also known as
fdk-aac-free). So, stop making fdk-aac optional.
This also eases up debugging audio issues, as it was previously not
clear, whether a build was compiled with fdk-aac, or not.