rdp: Add dedicated class for session metrics
The current session metrics implementation in the session-rdp class has
two issues:
1. The surface metrics are global and not per surface
2. It is too much entangled in the rendering handling

Issue 1 is problematic in the way, that it makes precise measurements
impossible in multi-monitor setups, as the skipped frames value and the
surface related timestamps are shared across all surfaces.

Issue 2 is problematic in the way, that it makes it hard to move the
rendering handling out of the session. Moving the rendering handling out
of the session would though make it easier to split up the rendering
behaviour to parallelize actions, that can be parallelized (something
that can also be useful for multi-monitor setups).

So, create a dedicated class for the session metrics. There, surface
metrics are per surface individually saved.
Also add a function to notify a layout change. In case the client
directly submits a new monitor config after session initialization, this
allows the session metrics class to reset old surface metrics, that may
already be saved.