rdp-surface: Also move refresh rate to surface renderer
While the surface renderer does not need the refresh rate itself, the
refresh rate is closer related to the rendering action, rather than the
description of the surface area:

The refresh rate is currently needed for two parts in the RDP backend:
1. The target rate of the PipeWire stream
2. The throttling threshold in the frame controller

Both parts, but mostly the latter one are more closely related to the
rendering behaviour. So, move the refresh rate to the surface renderer.