rdp-graphics-pipeline: Add API to submit frames
Currently, all rendering actions happen in the RDP graphics pipeline.
This was a relatively easy implementation. However, it does not provide
a lot of flexibility, which will be needed, when trying to render
multiple frames at the same time (for example, when each frame uses
different hardware resources at the same time).

This will change in future commits. The graphics pipeline won't render
any frames anymore. It will just submit any rendered frame.
For this, add a general function to submit a frame.
Not every frame is the same, some are AVC420 frames, some are AVC444v2
frames, some are RFX Progressive (Calista Progressive) encoded frames.

So, add handling for all of them. Use the previously added getter from
the render context to determine, which codec is used.