session-rdp: Combine related colour depth checks
When the client side supports codecs or the graphics pipeline, it also
needs to support 32-bit as colour depth.
Also, gnome-remote-desktop prefers the usage of the graphics pipeline
and codecs. But these require the usage of 32-bit as colour depth, which
needs to be written into the Bitmap Capability Set, that is sent by the
server side.
Both checks are closely related. So, combine their premise (the support
of codecs or the graphics pipeline) into one check.