rdp: Introduce render context
Currently, the surface renderer has to perform render permission checks,
that are specific to all rendering actions and not just the current
surface.
In addition to that, the current rendering logic only allows one
rendering operation at a time. For multi-monitor setups, and hardware
accelerated setups, this might not be the most efficient way to handle
rendering operations, especially when they can be parallelized.

To be able to take care of both issues, introduce a render context.
For now, this is just a plain GObject, but in the future it will contain
all surface rendering related items, that are specific to the current
monitor and to the current monitor layout.
Successfully acquiring this render context means, that rendering in
general is not inhibited or suppressed. Trying to acquire such render
context, implicitly moves the render permission logic for the whole
session to the renderer, this solves case 1.

In addition to that, add a method to release acquired render contexts.
This method will become more relevant in the future, as the idea here is
that the render context is acquired until the current rendering
operation for a given surface is completed.
Inhibiting all rendering actions would then mean, that new rendering
actions won't be able to occur, and that the end of the inhibition is
signalled, when all currently acquired rendering actions are released.
To signal the completion of the inhibition operation, add a signal to
the renderer, which is emitted, when the inhibition completed.

When a render context is released, do not clear it. In the future, the
render context is intended to carry specific render items for the given
surface. This can be a GFX surface for the given RDP surface, or
possibly an NVENC or VAAPI session for the given RDP surface.