rdp: Move output suppression handling from session to renderer
Moving the output suppression handling from session-rdp to the renderer
is relatively straight forward.
Also move the check for the suppression status into the surface
renderer, instead of leaving it in session-rdp.

In future commits, this permission check handling will be simplified and
be turned into one check, when further permission checks were moved.
For now, however, leave it just as a getter operation for a simple and
understandable transition.