build: Bump FreeRDP version requirement
When writing a DRDYNVC Data PDU, the server side and client side need to
ensure, that each Data PDU does not exceed 1600 Bytes.
If the DRDYNVC Data PDU exceeds 1600 Bytes, the PDU needs to be split
into multiple parts, where each part does not exceed 1600 Bytes.
This is a process, that is done in FreeRDP, when sending a PDU of a DVC.
When each segment of the overall Data PDU is in the process of being
sent, then unrelated Data PDUs for the same DVC MUST NOT be transmitted.

In the graphics pipeline, gnome-remote-desktop exposes, however, a case,
where exactly this situation can happen:
When gnome-remote-desktop sends a large PDU (e.g. a WireToSurface PDU)
and a Cache Import Offer PDU is received in the meantime, then it is
possible, that gnome-remote-desktop sends the Cache Import Reply PDU,
while the large PDU (e.g. a WireToSurface PDU) is sent.
The client can then not correctly assemble the actual Data PDU again.

FreeRDP-3.1.0 contains a fix ([0]), which handles that case with a
mutex. So, bump the FreeRDP version requirement to ensure, that
gnome-remote-desktop does not expose this bug anymore.

[0]: https://github.com/FreeRDP/FreeRDP/commit/6c26980585f862779f524592911144a6a55f5cfa