rdp: Move graphics reset handling to renderer
Moving the graphics reset handling is not that simple, compared to the
previous commit, because it partially involves the graphics pipeline
reset handling too.

First, add notify_new_desktop_layout() as method to the renderer. This
function works quite like notify_new_desktop_size() in session-rdp.
Instead of using general session flags to indicate a pending graphics
reset, use a normal boolean variable.

This first step now implicitly (due to the GFX related graphics reset
flag) requires that the notify_graphics_pipeline_ready()- and
notify_graphics_pipeline_reset()-methods are moved too.
The same applies to the reset_graphics() method itself.

To simplify the graphics reset handling change it however:
Currently, the layout manager has a dependency on the rendering
handling, as it needs to inhibit any rendering actions, before
attempting to make any layout changes.
At the same time, however, the rendering handling needs to know the
monitor layout for the graphics reset handling.
To get rid of this circular dependency, set the monitor layout in the
RDP context, when changing the layout.
When the graphics reset happens, read the monitor layout from the RDP
context.
This also simplifies the API of the layout manager.

With this simplification, the layout manager still has a dependency on
the renderer, but not vice-versa.