meson: Install a sysusers conf file to automatically create the grd user
The daemon-system is run as gnome-remote-desktop user. Use the
systemd-sysusers utility to create that user.
Also use the systemd-tmpfiles to set the ownership of the conf file.

A home directory is also needed when the GKeyFile backend is used to
store the credentials. So use the systemd-tmpfiles utility to create it.