routing-token: On peeking, check also if RDSTLS is requested
The routing token is contained in the X.224 Connection Request PDU.
In that PDU there's also the RDP Negotiation Request field struct which
contains the requestedProtocols field (the security protocols the RDP
client supports).

The whole X.224 Connection Request PDU is already peeked. So it is just
needed to find, whether the requestedProtocols exists and then check if
the RDSTLS flag value is set.

The rdp-server now also exposes, whether RDSLTS is requested, when
emitting incoming-redirected-connection signal.