daemon-system: Add use_system_credentials arg to take-client-ready
This parameter is TRUE, when the client is mstsc and it isn't
advertising support for RDSTLS. In that case mstsc reconnects using the
same system credentials instead of the one-time credentials sent in the
Server Redirection PDU.

Only on the first connection of the client, it can be known if it's
mstsc (on post-connected), because in redirected connections the
daemon-system only peeks the Client X.224 Connection Request PDU and
then handovers it to a daemon-handover.
On the other side, knowing whether RDSTLS is requested happens in
redirected connections when peeking the Client X.224 Connection Request
PDU (for more info see previous commits).