daemon-system: Add GetSystemCredentials method to handover iface
Add a new method called GetSystemCredentials to the handover interface,
which will be used by the daemon-handover, when it receives the
TakeClientReady signal with the use_system_credentials field set to
TRUE.

Check that this daemon-handover is allowed to get those credentials, and
after giving them these, set use_system_credentials to FALSE to not
always expose them.
use_system_credentials is always set when the client is redirected.