daemon-handover: Fix clearing logout timeout
There can be a situation, where the client is handovered properly but
fails on authentication. In that case, the logout timeout source was
already cleared so the session would be orphaned without any RDP client
displaying it.

Fix this by clearing the logout timeout source when the client
successfully authenticated (at the new-incoming-connection signal
handler).