prompt: Make it more generic to allow reusing it
Using the new struct GrdPromptDefinition, it is possible to configure
the summary, body, accept_label and cancel_label.

Set the urgency level of the notification to critical to ensure it
doesn't disappear. This fixes an issue when the notification disappears
without being able to handle an answer.

Show an accept button or cancel button depending on, whether
accept_label or cancel_label is set.